update notebook image display behavior
