+++ libs/components/views/Showcase/components/notebook-image.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React from 'react'
 import Image from 'next/image'
 import { ZoomIn } from 'lucide-react'
 import { cn } from '../../../../utils/cn'
@@ -7,23 +7,18 @@ import { hexToDataUrl } from '../../../../utils/rgb-to-data-url'
 export function NotebookImage({
   serial_num,
   mark_name,
-  onError,
   className,
-  isSlider
+  isSlider,
+  has_icon,
+  lazy = true
 }: {
   serial_num: string
   mark_name: string
-  onError?: () => void
   className?: string
   isSlider?: boolean
+  has_icon: boolean
+  lazy?: boolean
 }) {
-  const [isError, setIsError] = useState(false)
-
-  function handleImageError() {
-    setIsError(true)
-    onError?.()
-  }
-
   return (
     <div
       className={cn(
@@ -31,30 +26,32 @@ export function NotebookImage({
         className
       )}
     >
-      <Image
-        onError={handleImageError}
-        width={1200}
-        loading="lazy"
-        height={1200}
-        src={
-          isError
-            ? '/assets/icons/notebook-icon.svg'
-            : `/media/img/${serial_num}/icon.jpg`
-        }
-        alt={`${mark_name} notebook image`}
-        placeholder="blur"
-        blurDataURL={hexToDataUrl('#F3F6FB')}
-        className={cn(
-          'w-full transition-all duration-300 rounded-lg',
-          !isError && isSlider && 'group-hover:bg-white group-hover:opacity-50'
-        )}
-      />
+      <div className="relative aspect-square w-full h-full">
+        <Image
+          fill
+          src={
+            !has_icon
+              ? '/assets/icons/notebook-icon.svg'
+              : `/media/img/${serial_num}/icon.jpg`
+          }
+          alt={`${mark_name} notebook image`}
+          placeholder="blur"
+          blurDataURL={hexToDataUrl('#F3F6FB')}
+          loading={lazy ? 'lazy' : 'eager'}
+          className={cn(
+            'w-full transition-all duration-300 rounded-lg',
+            has_icon &&
+              isSlider &&
+              'group-hover:bg-white group-hover:opacity-50'
+          )}
+        />
+      </div>
 
       {isSlider && (
         <ZoomIn
           className={cn(
             'absolute text-primary top-1/2 left-1/2 -translate-y-1/2 -translate-x-1/2 hidden transition-all duration-300',
-            !isError && 'group-hover:block'
+            has_icon && 'group-hover:block'
           )}
         />
       )}
